refactor(web): decompose AppSidebar and remove shallow containers
Extract useAreaProjectTree and useSidebarDialogs hooks from the
346-line AppSidebar, separating data-fetching and dialog-lifecycle
concerns from layout. Remove single-caller Item/NewItem container
modules that were pure pass-throughs.

Multi-caller containers (ItemRowContainer, ProcessItemDialogContainer)
are kept — they prevent duplication across screens and dashboard.
