--- apps/web/src/features/items/new-item/new-item-dialog-container.tsx
@@ -1,25 +1,0 @@
-import { useCreateItem } from "@/features/items/use-create-item";
-import { NewItemDialog } from "./new-item-dialog";
-
-interface NewItemDialogProps {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-}
-
-export function NewItemDialogContainer({
-  open,
-  onOpenChange,
-}: NewItemDialogProps) {
-  const createItem = useCreateItem();
-
-  return (
-    <NewItemDialog
-      open={open}
-      onOpenChange={onOpenChange}
-      onSubmit={async (value) => {
-        await createItem(value);
-        onOpenChange(false);
-      }}
-    />
-  );
-}
